move spy_builtin and the code to parse signatures to its own module, vm.sig
